+++ lib/https.js
@@ -41,7 +41,7 @@ function Server(opts, requestListener) {
   }
 
   this.addListener('clientError', function(err, conn) {
-    conn.destroy(err);
+    conn.destroy();
   });
 
   this.timeout = 2 * 60 * 1000;
