+++ test/simple/test-http-url.parse-https.request.js
@@ -37,7 +37,7 @@ testURL.rejectUnauthorized = false;
 
 function check(request) {
   // assert that I'm https
-  assert.ok(request.socket.encrypted);
+  assert.ok(request.socket._secureEstablished);
 }
 
 var server = https.createServer(httpsOptions, function(request, response) {
