+++ test/simple/test-https-localaddress.js
@@ -35,8 +35,8 @@ var options = {
 };
 
 var server = https.createServer(options, function (req, res) {
-  console.log("Connect from: " + req.connection.socket.remoteAddress);
-  assert.equal('127.0.0.2', req.connection.socket.remoteAddress);
+  console.log("Connect from: " + req.connection.remoteAddress);
+  assert.equal('127.0.0.2', req.connection.remoteAddress);
 
   req.on('end', function() {
     res.writeHead(200, { 'Content-Type': 'text/plain' });
