--- test/simple/test-tls-server-slab.js
@@ -1,66 +1,0 @@
-// Copyright Joyent, Inc. and other Node contributors.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a
-// copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to permit
-// persons to whom the Software is furnished to do so, subject to the
-// following conditions:
-//
-// The above copyright notice and this permission notice shall be included
-// in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
-// NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
-// OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
-// USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-var common = require('../common');
-var assert = require('assert');
-var tls = require('tls');
-var fs = require('fs');
-
-var clientConnected = 0;
-var serverConnected = 0;
-
-var options = {
-  key: fs.readFileSync(common.fixturesDir + '/keys/agent1-key.pem'),
-  cert: fs.readFileSync(common.fixturesDir + '/keys/agent1-cert.pem')
-};
-
-tls.SLAB_BUFFER_SIZE = 100 * 1024;
-
-var server = tls.Server(options, function(socket) {
-  assert(socket._buffer.pool.length == tls.SLAB_BUFFER_SIZE);
-  if (++serverConnected === 2) {
-    server.close();
-  }
-});
-
-server.listen(common.PORT, function() {
-  var client1 = tls.connect({
-    port: common.PORT,
-    rejectUnauthorized: false
-  }, function() {
-    ++clientConnected;
-    client1.end();
-  });
-
-  var client2 = tls.connect({
-    port: common.PORT,
-    rejectUnauthorized: false
-  });
-  client2.on('secureConnect', function() {
-    ++clientConnected;
-    client2.end();
-  });
-});
-
-process.on('exit', function() {
-  assert.equal(clientConnected, 2);
-  assert.equal(serverConnected, 2);
-});
